Moves attack builders in a specific package; create new abstract class for Bruteforce builders to generalize methods and expose them with attack-specific terms
